+++ ExpandWorldPrefabs/Helper.cs
@@ -1,33 +1,32 @@
-
-
-using System.Linq;
-using ExpandWorldData;
+using System;
 
 namespace ExpandWorld.Prefab;
 
 public class Helper2
 {
 
-  public static Range<string> Range(string arg)
+  public static string ReplaceParameters(string str, string prefab, string parameter)
   {
-    var range = arg.Split('-').ToList();
-    if (range.Count > 1 && range[0] == "")
-    {
-      range[0] = "-" + range[1];
-      range.RemoveAt(1);
-    }
-    if (range.Count > 2 && range[1] == "")
-    {
-      range[1] = "-" + range[2];
-      range.RemoveAt(2);
-    }
-    if (range.Count == 1) return new(range[0]);
-    else return new(range[0], range[1]);
-
+    var split = parameter.Split(' ');
+    return str.Replace("$$prefab", prefab)
+      .Replace("$$par0", split.Length > 0 ? split[0] : "")
+      .Replace("$$par1", split.Length > 1 ? split[1] : "")
+      .Replace("$$par2", split.Length > 2 ? split[2] : "")
+      .Replace("$$par3", split.Length > 3 ? split[3] : "")
+      .Replace("$$par4", split.Length > 4 ? split[4] : "")
+      .Replace("$$par", parameter);
   }
-  public static Range<int> IntRange(string arg)
+  public static bool CheckWild(string wild, string str)
   {
-    var range = Range(arg);
-    return new(int.Parse(range.Min), int.Parse(range.Max));
+    if (wild == "*")
+      return true;
+    if (wild[0] == '*' && wild[wild.Length - 1] == '*')
+      return str.ToLowerInvariant().Contains(wild.Substring(1, wild.Length - 2).ToLowerInvariant());
+    if (wild[0] == '*')
+      return str.EndsWith(wild.Substring(1), StringComparison.OrdinalIgnoreCase);
+    else if (wild[wild.Length - 1] == '*')
+      return str.StartsWith(wild.Substring(0, wild.Length - 1), StringComparison.OrdinalIgnoreCase);
+    else
+      return str.Equals(wild, StringComparison.OrdinalIgnoreCase);
   }
 }